backend/requirements.txt
Flask==3.1.0
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.7.1
Flask-Bcrypt==1.0.1
Flask-Migrate==4.1.0
Flask-CORS==5.0.1
Flask-RESTful==0.3.10
PyMySQL==1.1.1
cryptography==44.0.0
python-dotenv==1.0.1
